refactor: url structure URL 구조 개선
