--- packages/front-end/app/dashboard/layout.tsx
@@ -1,23 +1,0 @@
-import { css } from "@/styled-system/css";
-import { ReactNode } from "react";
-import Sidebar from "./_components/Sidebar";
-
-interface PropType {
-  children: ReactNode;
-}
-
-export default function Layout({ children }: PropType) {
-  return (
-    <>
-      <main className={css({ display: "flex", height: "100vh" })}>
-        <Sidebar />
-        <div className={css({
-          flexGrow: 1,
-          padding: "2em 1em",
-        })}>
-          {children}
-        </div>
-      </main>
-    </>
-  );
-}
