+++ packages/front-end/app/sessions/participant/_components/SessionHeader.tsx
