+++ packages/front-end/app/sessions/participant/_components/SessionTable.tsx
